sdcard partition: Handle single partition sdcard
When repartitioning SDEXT as userdata, the size of the sdcard was
indeterminate if SDEXT did not exist on startup. This patch fixes that
by using the Primary_Block_Device instead of the Actual_Block_Device
when the TW_DATA_ON_SDEXT feature is enabled.

Also, when creating the SDEXT partition, do not create a filesystem on
it until the wipe is done (use mkpart instead of mkpartfs). This speeds
up the process considerably as we don't have to wait for a large ext2 fs
to be created.

Change-Id: Ifc23c6bb8a07bc9031d307ae4d387547549cd633